feat: show actions when multiple nodes are selected
