+++ packages/frontend/editor-ui/src/components/FocusPanel.vue
@@ -37,6 +37,7 @@ import { useNDVStore } from '@/stores/ndv.store';
 import { useVueFlow } from '@vue-flow/core';
 import ExperimentalFocusPanelHeader from '@/components/canvas/experimental/components/ExperimentalFocusPanelHeader.vue';
 import { useTelemetryContext } from '@/composables/useTelemetryContext';
+import { ContextMenuAction } from '@/composables/useContextMenuItems';
 
 defineOptions({ name: 'FocusPanel' });
 
@@ -47,6 +48,7 @@ const props = defineProps<{
 const emit = defineEmits<{
 	focus: [];
 	saveKeyboardShortcut: [event: KeyboardEvent];
+	contextMenuAction: [action: ContextMenuAction, nodeIds: string[]];
 }>();
 
 // ESLint: false positive
@@ -573,6 +575,7 @@ function onOpenNdv() {
 					:node="node"
 					:nodes="vueFlow.getSelectedNodes.value"
 					@open-ndv="onOpenNdv"
+					@context-menu-action="(action, nodeIds) => emit('contextMenuAction', action, nodeIds)"
 				/>
 				<div v-else :class="[$style.content, $style.emptyContent]">
 					<div :class="$style.emptyText">
