+++ packages/frontend/editor-ui/src/components/canvas/experimental/components/ExperimentalNodeDetailsDrawer.vue
@@ -2,30 +2,39 @@
 import { useExpressionResolveCtx } from '@/components/canvas/experimental/composables/useExpressionResolveCtx';
 import { ExpressionLocalResolveContextSymbol } from '@/constants';
 import { type INodeUi } from '@/Interface';
-import { N8nButton, N8nText } from '@n8n/design-system';
+import { N8nButton, N8nKeyboardShortcut, N8nText } from '@n8n/design-system';
 import { type GraphNode } from '@vue-flow/core';
 import { computed, provide } from 'vue';
 import ExperimentalCanvasNodeSettings from './ExperimentalCanvasNodeSettings.vue';
-import { useContextMenu } from '@/composables/useContextMenu';
+import { ContextMenuAction, useContextMenuItems } from '@/composables/useContextMenuItems';
 
 const { node, nodes } = defineProps<{ node: INodeUi; nodes: GraphNode[] }>();
 
-const emit = defineEmits<{ openNdv: [] }>();
+const emit = defineEmits<{
+	openNdv: [];
+	contextMenuAction: [ContextMenuAction, nodeIds: string[]];
+}>();
 
 const expressionResolveCtx = useExpressionResolveCtx(computed(() => node));
-const contextMenu = useContextMenu();
+const nodeIds = computed(() => nodes.map((n) => n.id));
+const contextMenuItems = useContextMenuItems(nodeIds);
 
 provide(ExpressionLocalResolveContextSymbol, expressionResolveCtx);
 </script>
 
 <template>
 	<div :class="$style.component">
-		<N8nText v-if="nodes.length > 1" color="text-base">
+		<N8nText v-if="nodes.length > 1" tag="div" color="text-base" :class="$style.multipleNodes">
 			<div>{{ nodes.length }} nodes selected</div>
-			<ul>
-				<li v-for="action of contextMenu.actions.value" :key="action.id">
-					<N8nButton text type="tertiary">
+			<ul :class="$style.multipleNodesActions">
+				<li v-for="action of contextMenuItems" :key="action.id" :class="$style.multipleNodesAction">
+					<N8nButton
+						type="secondary"
+						:disabled="action.disabled"
+						@click="emit('contextMenuAction', action.id, nodeIds)"
+					>
 						{{ action.label }}
+						<N8nKeyboardShortcut v-if="action.shortcut" v-bind="action.shortcut" />
 					</N8nButton>
 				</li>
 			</ul>
@@ -48,10 +57,50 @@ provide(ExpressionLocalResolveContextSymbol, expressionResolveCtx);
 
 <style lang="scss" module>
 .component {
-	display: flex;
-	align-items: center;
-	justify-content: center;
 	height: 100%;
 	overflow: auto;
+}
+
+.multipleNodes {
+	min-height: 100%;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: start;
+	padding: var(--spacing-3xl) var(--spacing-m);
+	gap: var(--spacing-m);
+}
+
+.multipleNodesActions {
+	align-self: stretch;
+	list-style-type: none;
+}
+
+.multipleNodesAction {
+	margin-top: -1px;
+
+	& button {
+		width: 100%;
+		border-radius: 0;
+		position: relative;
+		display: flex;
+		align-items: center;
+		justify-content: space-between;
+		border-color: var(--border-color-light) !important;
+	}
+
+	&:first-of-type button {
+		border-top-left-radius: var(--border-radius-base);
+		border-top-right-radius: var(--border-radius-base);
+	}
+
+	&:last-of-type button {
+		border-bottom-left-radius: var(--border-radius-base);
+		border-bottom-right-radius: var(--border-radius-base);
+	}
+
+	& button:hover {
+		z-index: 1;
+	}
 }
 </style>
